Updated README with course limitations
Added a new section in the README file to clarify what the course will not cover. This includes an explanation that not all AWS services will be discussed due to their vast number. Instead, the focus is on understanding how to write CloudFormation templates effectively, which can then be applied across any AWS service.