Updated EC2 instance to use IAM role
The EC2 instance has been updated to utilize an IAM role and Instance Profile. The IAM role is configured with a policy that allows it to assume the role, and the policy grants permissions for all EC2 actions. The Instance Profile is associated with this IAM role. Additionally, the ImageId of the EC2 instance has been updated, and metadata options have been added requiring HttpTokens.