Added new S3 bucket with specific access controls
A new AWS S3 bucket has been created with defined ownership and public access block configurations. The object ownership is set to ObjectWriter, allowing the writer of an object to own it. Public access block configurations are all set to false, meaning there are no restrictions on public access at this time.