Added Internet Gateway and Route Table to VPC
In this update, we've added an Internet Gateway and a Route Table to the Virtual Private Cloud (VPC). The Internet Gateway has been attached to the VPC. A route for all traffic (0.0.0.0/0) through the gateway has been set up in the Route Table. Additionally, we've associated our Subnet with this new Route Table.