Added new EC2 mapping and parameters
A new YAML file has been created that includes mappings for different AWS regions to their respective AMIs. It also introduces a parameter for environment name, which can be either 'development' or 'production'. The instance type varies based on the environment - 't2.micro' for development and 't2.small' for production. The resources section now includes an EC2 instance with properties set according to these mappings and parameters.