Added documentation for `Fn::GetAtt` function
The README.md file has been updated to include a description of the `Fn::GetAtt` function in AWS CloudFormation. This addition provides an explanation of how this function can be used to retrieve attribute values from resources within a template, such as the public IP address of an EC2 instance or the DNS name of an S3 bucket.