Added cfnlint to linter and created new EC2 instance with conditions
In this update, we've added the CloudFormation Linter (cfnlint) to our list of enabled linters. This will help us ensure that our AWS CloudFormation templates are accurate and secure.

We've also introduced a new YAML file for an EC2 instance. This includes a condition for creating a volume in production environments. The file outlines the necessary resources such as VPC, Subnet, Internet Gateway, Route Table, Security Group, Launch Template and EC2 Instance itself. It also includes conditions for creating resources specifically for production environments.