Added AWS CloudFormation template for EC2 instance
This commit introduces a new AWS CloudFormation YAML file that sets up an EC2 instance along with necessary network resources. The template includes parameters for the instance type, and creates resources such as VPC, Subnet, Internet Gateway, Route Table, Security Group and Launch Template. It also associates the subnet with the route table and attaches the internet gateway to the VPC. The security group is configured to allow SSH access from a trusted IP. An encrypted EBS volume is attached to the EC2 instance using KMS key ID. Outputs are provided for VPCId, SubnetId, SecurityGroupId and InstanceId.